+++ game_patch/misc/alpine_settings.h
@@ -152,6 +152,7 @@ struct AlpineGameSettings
     }
     bool show_glares = true;
     bool weather = true;
+    bool caustics = true;
     bool show_enemy_bullets = true;
     bool fps_counter = true;
     static constexpr int min_fps_counter_average_ms = 0;